refactor(userspace): extract zone/screen snapshots to protocol_zones.go
Third code-motion increment of the protocol.go wire-protocol split
(#4839). Moves the security-zone and screen-profile configuration
snapshots into a cohesive sibling protocol_zones.go in the same
userspace package.

Moved verbatim (no rename, no logic or json-tag change): ZoneSnapshot,
ScreenProfileSnapshot, ScreenMissingProfileRef.

Behavior byte-identical; json tags preserved for the Rust userspace-dp
deserializer. Build passes; gofmt clean.
